refactor drawing for lower memory use
To draw circles, I have to pick between:
1. bitmaptools.circle(), which always draws 1 px wide strokes
   with no option for a fill color
2. vectorio.Circle(), which always draws filled circles with
   weird looking 1 px spikes on the top, left, right, and bottom

What I want is thicker strokes with optional fill. To get that,
I use a chain of bitmaptools.circle(), bitmaptools.boundary_fill(),
and bitmaptools.blit(). Most of the complexity is so that I can
draw a stroke of the background color on top of the large ring to
separate the dots from the ring. To do that, I'm alpha blending
small dot bitmaps on top of the big background bitmap which has the
large 3px wide ring. Previously I was doing alpha blending using
two large displayio.Bitmap objects with a transparent palette
color in the upper one.

This commit refactors how the alpha blending works to save a bunch
of memory by using one large bitmap and two small ones (dot blit
sources).